+++ app/src/main/java/com/episode6/headachetracker/HeadacheTrackerApp.kt
@@ -3,6 +3,7 @@ package com.episode6.headachetracker
 import android.app.Application
 import android.content.Context
 import com.episode6.headachetracker.data.AutoExportManager
+import com.episode6.headachetracker.data.SecondPillReminderWorker
 import com.episode6.headachetracker.di.AppGraph
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.createGraphFactory
@@ -18,6 +19,7 @@ class HeadacheTrackerApp : Application() {
         appGraph = createGraphFactory<AppGraph.Factory>().create(this)
         appGraph.inject(this)
         autoExportManager.startObserving()
+        SecondPillReminderWorker.ensureNotificationChannel(this)
     }
 }